requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
transformers==4.35.2
torch==2.1.1
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
newsapi-python==0.2.7
feedparser==6.0.10
python-dotenv==1.0.0
sqlite3-utils==3.35
textblob==0.17.1
vaderSentiment==3.3.2
scikit-learn==1.3.2
wordcloud==1.9.2
nltk==3.8.1